pyproject.toml
[project]
name = "robottraining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gymnasium>=0.29",
    "imageio>=2.35",
    "imageio-ffmpeg>=0.5",
    "mujoco>=3.1",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "stable-baselines3>=2.3",
    "torch>=2.2",
    "tensorboard>=2.17",
    "tqdm>=4.66",
    "pytest>=8",
]

[tool.pytest.ini_options]
pythonpath = ["."]
